refactor: Calendar get tile props
